+++ ssr/uno.ts
@@ -4,12 +4,21 @@
  * (https://github.com/dragonwocky/nadder) under the MIT license
  */
 
-import { initUno, modernNormalize, unoPreset, unoTypography } from "../deps.ts";
+import {
+  modernNormalize,
+  unoGenerator,
+  unoPreset,
+  unoTypography,
+} from "../deps.ts";
 import { reduceTemplate } from "../util.ts";
 
-const expandGroups = (className: string) => {
+const expandUtilityGroups = (
+  t: string | TemplateStringsArray | string[],
+  ...s: unknown[]
+) => {
   // e.g. font-(bold mono) m(y-4 x-3) dark:(font-blue hover:(p-8 h-full))
   // returns: font-bold font-mono my-4 mx-3 dark:font-blue dark:hover:p-8 dark:hover:h-full
+  let className = typeof t === "string" ? t : reduceTemplate(t, ...s);
   const pattern = /([^\s'"`;>=]+)\(([^(]+?)\)/g,
     replace = () => {
       className = className.replaceAll(
@@ -23,27 +32,19 @@ const expandGroups = (className: string) => {
   return className;
 };
 
-type UnoConfig = Parameters<typeof initUno>[0];
-type UnoGenerator = ReturnType<typeof initUno>;
+type UnoConfig = Parameters<typeof unoGenerator>[0];
+type UnoGenerator = ReturnType<typeof unoGenerator>;
 type GenerateOptions = Parameters<UnoGenerator["generate"]>[1];
 
-const unoInstance = (config: UnoConfig = {}) => {
-  let cache = "", engine: undefined | UnoGenerator = undefined;
+const createUnoGenerator = (config: UnoConfig = {}) => {
   config.presets = config.presets ??
     [unoPreset({ dark: "class", variablePrefix: "uno-" }), unoTypography()];
   config.preflights = config.preflights ?? [{ getCSS: () => modernNormalize }];
-  return {
-    uno: (t: TemplateStringsArray | string[], ...s: unknown[]) => {
-      const className = expandGroups(reduceTemplate(t, ...s));
-      cache += ` ${className}`;
-      return className;
-    },
-    css: async (options: GenerateOptions = { minify: true }) => {
-      if (!engine) engine = initUno(config);
-      const { css } = await engine.generate(cache, options);
-      return css;
-    },
-  };
+  const engine = unoGenerator(config);
+  return async (
+    classList: string,
+    options: GenerateOptions = { minify: true },
+  ) => (await engine.generate(classList, options)).css;
 };
 
-export { unoInstance };
+export { createUnoGenerator, expandUtilityGroups };